perf(argocd): ignore resource update to reduce CPU usage
From https://argo-cd.readthedocs.io/en/stable/operator-manual/reconcile

By default, an Argo CD Application is refreshed every time a resource that
belongs to it changes.

Kubernetes controllers often update the resources they watch periodically,
causing continuous reconcile operation on the Application and a high CPU
usage on the argocd-application-controller

When a resource update is ignored, if the resource's health status does
not change, the Application that this resource belongs to will not be
reconciled.